backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
nba_api==1.2.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
tweepy==4.14.0
scikit-learn==1.3.2
tensorflow==2.14.0
transformers==4.35.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2
